requirements.txt
Flask==2.3.3
pandas==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
python-magic==0.4.27